\name{barcelos}
\alias{barcelos}
\docType{data}
\title{
30-day maxima rainfall at Barcelos Station
}
\description{
These data are the 30-day maxima rainfall at Barcelos Station, in Portugal, from 1931 to 2008. The data are contained in a numeric vector
}
\usage{data(barcelos)}
\format{
A numeric vector containing 918 observations.
}
\examples{
data(barcelos)
hist(barcelos, main=NULL)
}
\keyword{datasets}
