\name{fajardo}
\alias{fajardo}
\docType{data}
\title{
15-day maxima river food at Fajardo River
}
\description{
These data are the 15-day maxima river food at Fajardo River, in Porto Rico, from 1967 to 2008. The data are contained in a numeric vector
}
\usage{data(fajardo)}
\format{
A numeric vector containing 864 observations.
}
\examples{
data(fajardo)
hist(fajardo)
}
\keyword{datasets}
